+++ R/ctl_new_pillar.R
@@ -30,7 +30,34 @@
 #' @param title The title, derived from the name of the column in the data
 #'
 #' @export
+#' @examplesIf requireNamespace("palmerpenguins", quietly = TRUE) && requireNamespace("tibble", quietly = TRUE)
+#' # Create pillar objects
+#' ctl_new_pillar(
+#'   palmerpenguins::penguins,
+#'   palmerpenguins::penguins$species[1:3], width = 60
+#' )
+#' ctl_new_pillar(
+#'   palmerpenguins::penguins,
+#'   palmerpenguins::penguins$bill_length_mm[1:3],
+#'   width = 60
+#' )
+#'
+#' # Packed data frame
+#' ctl_new_compound_pillar(
+#'   tibble::tibble(),
+#'   palmerpenguins::penguins,
+#'   width = 60
+#' )
+#'
+#' # Packed matrix
+#' ctl_new_compound_pillar(tibble::tibble(), matrix(1:6, ncol = 2), width = 60)
+#'
+#' # Packed array
+#' ctl_new_compound_pillar(tibble::tibble(), Titanic, width = 60)
+#'
 #' @examples
+#'
+#' # Customize output
 #' lines <- function(char = "-") {
 #'   stopifnot(nchar(char) == 1)
 #'   structure(char, class = "lines")