+++ tests/testthat/setup.R
@@ -1,5 +1,5 @@
 rlang::with_options(
-  crayon.enabled = FALSE, crayon.colors = 1L, cli.num_colors = 1L,
+  cli.num_colors = 1L,
   {
     num_colors(forget = TRUE)
   }